Add sbt project for scala rewrite
