+++ rosette-scala/.scalafmt.conf
@@ -1,0 +1,6 @@
+project {
+  git = true
+  excludeFilters = ["target/"]
+}
+style = default
+rewrite.rules = [PreferCurlyFors, RedundantBraces, SortImports]
