+++ rosette-scala/project/Dependencies.scala
@@ -1,0 +1,5 @@
+import sbt._
+
+object Dependencies {
+  lazy val scalaTest = "org.scalatest" %% "scalatest" % "3.0.1"
+}
