+++ rosette-scala/project/plugins.sbt
@@ -1,0 +1,1 @@
+addSbtPlugin("com.lucidchart" % "sbt-scalafmt" % "1.9")
