+++ rosette-scala/src/test/scala/coop/rchain/rosette/HelloSpec.scala
@@ -1,0 +1,9 @@
+package coop.rchain.rosette
+
+import org.scalatest._
+
+class HelloSpec extends FlatSpec with Matchers {
+  "The Hello object" should "say hello" in {
+    "hello" shouldEqual "hello"
+  }
+}
